Make header video load last
The background video in the header is one of the larger files, and it loading basically whenever could contribute to the site being more sluggish. The video now is not downloaded by default, only the metadata is, until the script tells it to play after the rest of the page is loaded.