Purge CSS during generation
The CSS framework was embedded in the index page making it around 1.2MB, purgecss removes any unused CSS reducing it to a couple of KB. The downside is separateCSS files for each page. The following site has some suggestions for merging them back in: https://www.purgecss.com/guides/nuxt.